- Keep master's kk_op_contains / kk_collection_isEmpty wiring for Collection.contains/isEmpty.

- Remove stale kk_collection_contains runtime bridge and ABI spec.

- Restore __kk_string_builder_new_capacity_checked and fix __kk_string_builder_append_range to use existing UTF-16 helpers.

- Drop duplicate isSourceBackedArrayCopyCall definition.

+++ Scripts/diff_cases/uint_range.kt
@@ -1,4 +1,3 @@
-// SKIP-DIFF (DEBT-DIFF-007): surfaced by compile-exit parity fix; triage and split or fix before re-enabling
 fun main() {
     // Basic UIntRange creation and contains
     val range = 1u..10u
@@ -36,7 +35,7 @@ fun main() {
     (1u..5u).forEach { print("$it ") }
     println()
     println((1u..5u).map { it * 2u })
-    println((1u..5u).mapIndexed { index, value -> index + value })
+    println((1u..5u).mapIndexed { index, value -> index.toUInt() + value })
 
     // for loop
     for (i in 1u..5u) print("$i ")
